+++ db/migrate/20170509115010_add_initial_signatures_goal_to_petition_details.rb
@@ -1,0 +1,5 @@
+class AddInitialSignaturesGoalToPetitionDetails < ActiveRecord::Migration
+  def change
+    add_column :petition_plugin_details, :initial_signatures_goal, :integer, null: false, default: 10_000
+  end
+end